Fix RNN crash when operating on doubles.
RNN doesn't seem to like operating on doubles, so force input to float.

+++ src/librl/nn/core/rnn.py
@@ -87,7 +87,7 @@ class RecurrentKernel(nn.Module):
         else: cell_state = self.cell_state
 
         h0 = hidden_state.to(input.device), cell_state.to(input.device)
-        output, h1 = self.recurrent_layer(input, h0)
+        output, h1 = self.recurrent_layer(input.float(), h0)
 
         self.hidden_state, self.cell_state = h1
         # We really dont care about the progress / history of our state data.
